Path Planner Update
DFS and BSF are now one line of code.
The both call a generic graph search function.
DFS passes a stack as its argument to the generic search function.
BSF passes a queue as its argument to the generic search function.
With some work, A* can also become a one-liner.