Fixed graph initialization issue
Initially the graph was initialized with a default orientation of 0 which really screwed stuff up with the mapping before the position estimate was able to converge. This commit uses the orientation from the imu as the initial heading and fixes the problems when the robot initially at a different heading.