GPS Coordinates Fix (#875)
* corrected hector gps plugin for jessi and swervi

Realized that there was a problem with the hector gazebo plugin which was producing the wrong latitudes and longitudes in the simulation. The wrong behavior was that along the +X axis (East) latitude increased and along the +Y axis (North) longitude decreased. Instead we needed for +X to increase longitude and +Y to increase latitude. Thus I had to set the reference_heading in the jessii.urdf.xacro file and the swervi.urdf.xacro file to be 90 to fix this issue. Weird how its been going on for so long. Surprised nobody caught this for so long.

* Set magnetometer's orientation offset to zero so its yaw is zero when facing east

* Fixed all waypoints that were wrong due to hector plugin bug

* fixed magnetometer yaw offset for swervi model as well

* Changed IMU heading defaults

Typically most IMUs report 0 when facing north. Tried looking at the datasheet for the IMU we currently have (Hillcrest Labs FSM-9) but could not find that parameter. Probably will find it out during testing.

* updated yaw offset for localization node